add geolocation hook
